+++ lib/src/package_analyzer.dart
@@ -23,7 +23,6 @@ import 'platform.dart';
 import 'pubspec.dart';
 import 'sdk_env.dart';
 import 'utils.dart';
-import 'version.dart';
 
 class InspectOptions {
   final bool deleteTemporaryDirectory;
@@ -91,7 +90,7 @@ class PackageAnalyzer {
   }
 
   Future<Summary> _inspect(String pkgDir, InspectOptions options) async {
-    log.info("SDK: ${_toolEnv.dartSdkInfo.version}");
+    log.info("SDK: ${_toolEnv.runtimeInfo.sdkVersion}");
     if (_toolEnv.pubCacheDir != null) {
       log.fine("Using .package-cache: ${_toolEnv.pubCacheDir}");
     }
@@ -307,11 +306,6 @@ class PackageAnalyzer {
     }
     dartFileSuggestions.sort();
 
-    Map<String, Object> flutterVersion;
-    if (usesFlutter) {
-      flutterVersion = await _toolEnv.getFlutterVersion();
-    }
-
     DartPlatform platform;
     if (pkgPlatformConflict != null) {
       platform = new DartPlatform.conflict(
@@ -341,8 +335,7 @@ class PackageAnalyzer {
     suggestions.sort();
 
     return new Summary(
-      panaPkgVersion,
-      _toolEnv.dartSdkInfo.version,
+      _toolEnv.runtimeInfo,
       pubspec.name,
       pubspec.version,
       pubspec,
@@ -353,7 +346,6 @@ class PackageAnalyzer {
       pkgFitness,
       maintenance,
       suggestions.isEmpty ? null : suggestions,
-      flutterVersion: flutterVersion,
     );
   }